requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.txt
#
django-enumfields==0.10.1
django==2.1.2
djangorestframework==3.9.0
psycopg2-binary==2.7.5
pytz==2018.6
